// src/models/IoK8sApiAppsV1StatefulSetUpdateStrategy/types.ts
import { IoK8sApiAppsV1RollingUpdateStatefulSetStrategyModelOptions as IoK8sApiAppsV1RollingUpdateStatefulSetStrategy } from '../IoK8sApiAppsV1RollingUpdateStatefulSetStrategy';

export interface IoK8sApiAppsV1StatefulSetUpdateStrategyModelOptions {
  rollingUpdate?: IoK8sApiAppsV1RollingUpdateStatefulSetStrategy;
  type?: string;
}